+++ projects/nae-example-app/src/app/doc/demo-title-config/content/1/demo-title-config-content1-case-view.component.ts
@@ -23,7 +23,6 @@ import {Subject} from 'rxjs';
 
 interface ExampleInjectedData extends InjectedTabbedCaseViewData {
     exampleUseCache: boolean;
-    loadFilter?: Filter;
 }
 
 const localAllowedNetsFactory = (factory: AllowedNetsServiceFactory) => {