+++ CHANGELOG.md
@@ -7,7 +7,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Full Changelog: [https://github.com/netgrif/components/commits/v6.3.0](https://github.com/netgrif/components/commits/v6.3.0)
 
-## [6.3.0](https://github.com/netgrif/components/releases/tag/v6.3.0) (2023-03-08)
+## [6.3.0](https://github.com/netgrif/components/releases/tag/v6.3.0) (2023-03-24)
 
 ### Fixed
 
@@ -33,9 +33,16 @@ Full Changelog: [https://github.com/netgrif/components/commits/v6.3.0](https://g
 ### Added
 
 - [NAE-1678] User impersonation
+- [NAE-1809] UserList field show selected users
 - [NAE-1835] Change password component
 - [NAE-1841] Idle timer component
 
+
+## [6.2.8](https://github.com/netgrif/components/releases/tag/v6.2.8) (2023-03-22)
+
+### Added
+
+- [NAE-1851] Save selected headers in filter
 
 ## [6.2.7](https://github.com/netgrif/components/releases/tag/v6.2.7) (2022-12-19)